+++ packages/framework/src/Model/Order/Withdrawal/WithdrawalSettingFacade.php
@@ -7,9 +7,8 @@ namespace Shopsys\FrameworkBundle\Model\Order\Withdrawal;
 use DateTimeInterface;
 use Shopsys\FrameworkBundle\Model\Order\Order;
 use Shopsys\FrameworkBundle\Model\Order\OrderUrlGenerator;
-use Shopsys\FrameworkBundle\Model\Order\Withdrawal\Exception\WithdrawalException;
 
-class WithdrawalFacade
+class WithdrawalSettingFacade
 {
     public const string VARIABLE_ORDER_DETAIL_URL = '{order_detail_url}';
     public const string VARIABLE_ORDER_NUMBER = '{order_number}';
@@ -17,13 +16,11 @@ class WithdrawalFacade
     /**
      * @param \Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalSetting $withdrawalSetting
      * @param \Shopsys\FrameworkBundle\Model\Order\OrderUrlGenerator $orderUrlGenerator
-     * @param \Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalChecker $withdrawalChecker
      * @param \Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalDeadlineCalculation $withdrawalDeadlineCalculation
      */
     public function __construct(
         protected readonly WithdrawalSetting $withdrawalSetting,
         protected readonly OrderUrlGenerator $orderUrlGenerator,
-        protected readonly WithdrawalChecker $withdrawalChecker,
         protected readonly WithdrawalDeadlineCalculation $withdrawalDeadlineCalculation,
     ) {
     }
@@ -37,21 +34,6 @@ class WithdrawalFacade
         $withdrawalInstructions = $this->withdrawalSetting->getInstructions($order->getDomainId());
 
         return $this->replaceVariables($order, $withdrawalInstructions);
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Order\Order $order
-     * @return bool
-     */
-    public function canRequestWithdrawal(Order $order): bool
-    {
-        try {
-            $this->withdrawalChecker->checkOrderWithdrawal($order);
-
-            return true;
-        } catch (WithdrawalException) {
-            return false;
-        }
     }
 
     /**
