+++ packages/frontend-api/src/Model/Resolver/Order/OrderResolverMap.php
@@ -8,19 +8,19 @@ use Overblog\GraphQLBundle\Resolver\ResolverMap;
 use Override;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Model\Order\Order;
-use Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalFacade;
 use Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalRequestFacade;
+use Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalSettingFacade;
 
 class OrderResolverMap extends ResolverMap
 {
     /**
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
-     * @param \Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalFacade $withdrawalFacade
+     * @param \Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalSettingFacade $withdrawalSettingFacade
      * @param \Shopsys\FrameworkBundle\Model\Order\Withdrawal\WithdrawalRequestFacade $withdrawalRequestFacade
      */
     public function __construct(
         protected readonly Domain $domain,
-        protected readonly WithdrawalFacade $withdrawalFacade,
+        protected readonly WithdrawalSettingFacade $withdrawalSettingFacade,
         protected readonly WithdrawalRequestFacade $withdrawalRequestFacade,
     ) {
     }
@@ -49,13 +49,13 @@ class OrderResolverMap extends ResolverMap
                     return $order->getItemsSortedWithRelatedItems();
                 },
                 'withdrawalInstructions' => function (Order $order) {
-                    return $this->withdrawalFacade->getWithdrawalInstructions($order);
+                    return $this->withdrawalSettingFacade->getWithdrawalInstructions($order);
                 },
                 'canRequestWithdrawal' => function (Order $order) {
-                    return $this->withdrawalFacade->canRequestWithdrawal($order);
+                    return $this->withdrawalRequestFacade->canRequestWithdrawal($order);
                 },
                 'withdrawalDeadline' => function (Order $order) {
-                    return $this->withdrawalFacade->getWithdrawalDeadline($order);
+                    return $this->withdrawalSettingFacade->getWithdrawalDeadline($order);
                 },
                 'withdrawalRequest' => function (Order $order) {
                     return $this->withdrawalRequestFacade->findByOrder($order);
